+++ packages/api-explorer/__tests__/CodeSample.test.jsx
@@ -13,7 +13,7 @@ const props = {
   language: 'node',
   setLanguage: () => {},
   oasUrl: 'https://example.com/openapi.json',
-  operation: new Operation({}, '/pet/{id}', 'get'),
+  operation: new Operation({}, '/pet/{petId}', 'get'),
 };
 
 describe('tabs', () => {
@@ -61,7 +61,7 @@ describe('code examples', () => {
     const docProps = {
       ...props,
       language: 'node-simple',
-      operation: new Operation({}, '/pets/{petId}', 'get'),
+      operation: new Operation({}, '/pet/{petId}', 'get'),
     };
 
     const codeSample = shallow(
@@ -85,7 +85,7 @@ describe('code examples', () => {
   it('should display custom examples over pre-filled examples', () => {
     const docProps = {
       ...props,
-      operation: new Operation({}, '/pet/{id}', 'get'),
+      operation: new Operation({}, '/pet/{petId}', 'get'),
       examples: [
         {
           language: 'javascript',
@@ -114,7 +114,7 @@ describe('code examples', () => {
   it('should display custom examples even if SAMPLES_ENABLED is false', () => {
     const docProps = {
       ...props,
-      operation: new Operation({}, '/pet/{id}', 'get'),
+      operation: new Operation({}, '/pet/{petId}', 'get'),
       examples: [
         {
           language: 'javascript',
@@ -142,7 +142,7 @@ describe('code examples', () => {
   it('should not error if no code given', () => {
     const docProps = {
       ...props,
-      operation: new Operation({}, '/pet/{id}', 'get'),
+      operation: new Operation({}, '/pet/{petId}', 'get'),
       examples: [
         {
           language: 'javascript',
@@ -169,7 +169,7 @@ describe('code examples', () => {
   it('should not error if language requested cannot be auto-generated', () => {
     const docProps = {
       ...props,
-      operation: new Operation({}, '/pet/{id}', 'get'),
+      operation: new Operation({}, '/pet/{petId}', 'get'),
       language: 'css',
     };
 
@@ -197,7 +197,7 @@ describe('code examples', () => {
   it('should not render sample if language is missing', () => {
     const docProps = {
       ...props,
-      operation: new Operation({}, '/pet/{id}', 'get'),
+      operation: new Operation({}, '/pet/{petId}', 'get'),
       examples: [
         {
           code: 'console.log(1);',
@@ -229,7 +229,7 @@ describe('code examples', () => {
   it('should render first of examples if language does not exist', () => {
     const docProps = {
       ...props,
-      operation: new Operation({}, '/pet/{id}', 'get'),
+      operation: new Operation({}, '/pet/{petId}', 'get'),
       examples: [
         {
           language: 'javascript',
@@ -278,10 +278,36 @@ describe('code examples', () => {
     expect(codeSample.find('.hub-lang-switch-node').text()).toBe('Node');
   });
 
+  // Skipped until https://github.com/readmeio/api-explorer/issues/965 is resolved.
+  it.skip('should check the operation level extensions first', () => {
+    const operationSamplesEnabled = new Operation({}, '/pet/{petId}', 'get');
+    operationSamplesEnabled[extensions.SAMPLES_ENABLED] = true;
+    const languages = ['node', 'curl'];
+
+    const codeSample = shallow(
+      <CodeSample
+        {...props}
+        oas={
+          new Oas({
+            [extensions.SAMPLES_ENABLED]: false,
+            [extensions.SAMPLES_LANGUAGES]: languages,
+            servers: [{ url: 'http://example.com' }],
+          })
+        }
+        operation={operationSamplesEnabled}
+      />
+    );
+
+    expect(codeSample.find('.hub-code-auto')).toHaveLength(1);
+    // We only render one language at a time
+    expect(codeSample.find('.hub-code-auto pre')).toHaveLength(1);
+    expect(codeSample.find('.hub-lang-switch-node').text()).toBe('Node');
+  });
+
   it('should not display more than one example block at a time', () => {
     const docProps = {
       ...props,
-      operation: new Operation({}, '/pet/{id}', 'get'),
+      operation: new Operation({}, '/pet/{petId}', 'get'),
       language: 'javascript',
       examples: [
         {