+++ packages/api-explorer/src/form-components/ArrayField.jsx
@@ -1,13 +1,10 @@
 const React = require('react');
 const PropTypes = require('prop-types');
-const extensions = require('@readme/oas-extensions');
 
-const BaseArrayField = require('@readme/react-jsonschema-form/lib/components/fields/ArrayField').default;
-const { getDefaultRegistry, retrieveSchema } = require('@readme/react-jsonschema-form/lib/utils');
+const BaseArrayField = require('@readme/oas-form/src/components/fields/ArrayField').default;
+const { getDefaultRegistry, retrieveSchema } = require('@readme/oas-form/src/utils');
 
-function createArrayField(oas) {
-  const explorerEnabled = oas[extensions.EXPLORER_ENABLED];
-
+function createArrayField(explorerEnabled) {
   function hasPrimitives(props) {
     const { schema, registry = getDefaultRegistry() } = props;
     const { rootSchema } = registry;
@@ -49,6 +46,7 @@ function createArrayField(oas) {
   }
 
   ArrayField.propTypes = {
+    explorerEnabled: PropTypes.bool,
     formData: PropTypes.any,
     uiSchema: PropTypes.shape({}).isRequired,
   };