+++ packages/api-explorer/src/form-components/FileWidget.jsx
@@ -1,10 +1,6 @@
-const extensions = require('@readme/oas-extensions');
+const FileWidget = require('@readme/oas-form/src/components/widgets/FileWidget').default;
 
-const FileWidget = require('@readme/react-jsonschema-form/lib/components/widgets/FileWidget').default;
-
-function createFileWidget(oas) {
-  const explorerEnabled = oas[extensions.EXPLORER_ENABLED];
-
+function createFileWidget(explorerEnabled) {
   // Return a function that renders null when the explorer is disabled
   if (!explorerEnabled) return () => null;